Develop

+++ backend/src/main/java/com/intiwasi/backend/repository/CategoriaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.intiwasi.backend.entity.Categoria;
 
 public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {
-    List<Categoria> findByEstado(Integer estado);
+    List<Categoria> findByEstado(Byte estado);
     boolean existsByNomCategoria(String nomCategoria);
     Optional<Categoria> findByNomCategoria(String nomCategoria);
 }
