+++ backend/src/main/java/com/intiwasi/backend/repository/ProveedorRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.intiwasi.backend.entity.Proveedor;
 public interface ProveedorRepository extends JpaRepository <Proveedor, Integer> {
-    List<Proveedor> findByEstado(Integer estado);
+    List<Proveedor> findByEstado(Byte estado);
     // Métodos nuevos para validar si existen registros duplicados
     boolean existsByRuc(String ruc);
     boolean existsByNomProveedor(String nomProveedor);
